# internal/api/handler/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CastMember:
    properties:
      character:
        type: string
      name:
        type: string
      profile_path:
        type: string
    type: object
...
  models.MediaItem:
    properties:
      audio_codec:
        type: string
      backdrop_path:
        type: string
      bundle_status:
        type: string
      cast:
        items:
          $ref: '#/definitions/models.CastMember'
        type: array
      created_at:
        type: string
      director:
        type: string
      duration:
        description: seconds
        type: number
      episode_number:
        type: integer
      extra_posters:
        items:
          type: string
        type: array
      file_path:
        type: string
      file_size:
        type: integer
      height:
        type: integer
      id:
        type: string
      library_id:
        type: string
      media_type:
        $ref: '#/definitions/models.MediaType'
      mpd_path:
        type: string
      overview:
        type: string
      poster_path:
        type: string
      season_number:
        type: integer
      source_fingerprint:
        type: string
      source_status:
        type: string
      sub_jobs:
        items:
          $ref: '#/definitions/models.TranscodeSubJob'
        type: array
      title:
        type: string
      tmdb_id:
        description: Enriched metadata (nullable)
        type: integer
      transcode_progress:
        type: number
      transcode_sizes:
        $ref: '#/definitions/models.TranscodeSizesInfo'
      transcode_status:
        allOf:
        - $ref: '#/definitions/models.TranscodeStatus'
        description: Transcode
      tv_season_id:
        type: string
      tv_show_id:
        description: TV episode fields (nullable; populated only for MediaTypeEpisode
          items)
        type: string
      tv_show_title:
        type: string
      updated_at:
        type: string
      video_codec:
        type: string
      width:
        type: integer
      year:
        type: integer
    type: object
  models.MediaType:
    enum:
    - movie
    - tvshow
    - episode
    - music
    type: string
    x-enum-varnames:
    - MediaTypeMovie
    - MediaTypeTVShow
    - MediaTypeEpisode
    - MediaTypeMusic
  models.RenditionSize:
    properties:
      resolution:
        description: e.g., "360p", "480p", etc.
        type: string
      size:
        description: Total size of all files in the rendition directory in bytes
        type: integer
    type: object
...
  models.TranscodeSizesInfo:
    properties:
      renditions:
        items:
          $ref: '#/definitions/models.RenditionSize'
        type: array
      total_size:
        type: integer
    type: object
  models.TranscodeStatus:
    enum:
    - none
    - pending
    - processing
    - done
    - failed
    type: string
    x-enum-varnames:
    - TranscodeStatusNone
    - TranscodeStatusPending
    - TranscodeStatusProcessing
    - TranscodeStatusDone
    - TranscodeStatusFailed
  models.TranscodeSubJob:
    properties:
      audio_bitrate_k:
        type: integer
      codec:
        type: string
      created_at:
        type: string
      error_msg:
        type: string
      finished_at:
        type: string
      height:
        type: integer
      id:
        type: string
      job_id:
        type: string
      media_item_id:
        type: string
      profile_id:
        type: string
      profile_name:
        type: string
      progress:
        description: 0-100
        type: number
      started_at:
        type: string
      status:
        $ref: '#/definitions/models.TranscodeStatus'
      type:
        description: '"video" or "subtitles"'
        type: string
      video_bitrate_k:
        type: integer
      width:
        type: integer
      worker_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST & WebSocket API specification for Prism, a self-hosted media server.
  title: Prism Media Server API
  version: "1.0"
paths:
...
  /movies:
    get:
      description: Lists media items (movies, episodes) with filtering/sorting capabilities.
      parameters:
      - description: Library ID
        format: uuid
        in: query
        name: library_id
        type: string
      - description: Media Type
        in: query
        name: media_type
        type: string
      - description: Sort order
        in: query
        name: sort
        type: string
      - default: 20
        description: Recent items count limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MediaItem'
            type: array
        "400":
          description: Invalid library ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthenticated
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List Media Items
      tags:
      - Media Items
  /movies/{id}:
    delete:
      parameters:
      - description: Media ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Media item deleted successfully
        "400":
          description: Invalid media ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (requires Admin status)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Media item not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Media Item (Admin Only)
      tags:
      - Media Items
    get:
      parameters:
      - description: Media ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MediaItem'
        "400":
          description: Invalid media ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Media item not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Media Item
      tags:
      - Media Items
  /movies/{id}/transcode-sizes:
    get:
      description: Returns the size of each resolution subdirectory and the total
        size in the media item's transcode bundle.
      parameters:
      - description: Media ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TranscodeSizesInfo'
        "400":
          description: Invalid media ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Media item not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Media Transcode Sizes
      tags:
      - Media Items
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <your-token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
  WorkerAuth:
    description: The secret API Key assigned to the transcode worker.
    in: header
    name: X-Worker-API-Key
    type: apiKey
swagger: "2.0"